demonstrating functions and nested loops in C

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,16 +10,16 @@
 int main(void)
 {
 	int count;
-	int num1 = 0;
-	int num2 = 1;
-	int sum = 0;
+	long int num1 = 0;
+	long int num2 = 1;
+	long int sum = 0;
 
 	for (count = 0; count < 50; count++)
 	{
 		sum = num1 + num2;
 		num1 = num2;
 		num2 = sum;
-		printf("%d", sum);
+		printf("%lu", sum);
 		if (count < 49)
 		{
 			printf(", ");
